File version delete when file doesn't exist

+++ web/concrete/src/File/Version.php
@@ -578,7 +578,9 @@ class Version
 
         $fsl = $this->getFile()->getFileStorageLocationObject()->getFileSystemObject();
         $fre = $this->getFileResource();
-        $fsl->delete($fre->getPath());
+        if ($fsl->has($fre->getPath())) {
+            $fsl->delete($fre->getPath());
+        }
     }
 
     /**
